chore. update api and module
