+++ .github/README.md
@@ -1,0 +1,7 @@
+# ![logo](https://raw.githubusercontent.com/azerothcore/azerothcore.github.io/master/images/logo-github.png) AzerothCore
+
+## mod-notify-muted
+
+- Latest build status with azerothcore:
+
+[![Build Status](https://github.com/azerothcore/mod-notify-muted/workflows/core-build/badge.svg?branch=master&event=push)](https://github.com/azerothcore/mod-notify-muted)
