+++ .github/workflows/core-build.yml
@@ -1,45 +1,12 @@
 name: core-build
 on:
   push:
+    branches:
+      - 'master'
   pull_request:
 
 jobs:
   build:
-    strategy:
-      fail-fast: false
-      matrix:
-        compiler: [clang]
-    runs-on: ubuntu-20.04
-    name: ${{ matrix.compiler }}
-    env:
-      COMPILER: ${{ matrix.compiler }}
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          repository: 'azerothcore/azerothcore-wotlk'
-          ref: 'master'
-          submodules: 'recursive'
-      - uses: actions/checkout@v2
-        with:
-          submodules: 'recursive'
-          path: 'modules/mod-notify-muted'
-      - name: Cache
-        uses: actions/cache@v1.1.2
-        with:
-          path: /home/runner/.ccache
-          key: ccache:${{ matrix.compiler }}:${{ github.ref }}:${{ github.sha }}
-          restore-keys: |
-            ccache:${{ matrix.compiler }}:${{ github.ref }}
-            ccache:${{ matrix.compiler }}
-      - name: Configure OS
-        run: source ./apps/ci/ci-install.sh
-        env:
-          CONTINUOUS_INTEGRATION: true
-      - name: Import db
-        run: source ./apps/ci/ci-import-db.sh
-      - name: Build
-        run: source ./apps/ci/ci-compile.sh
-      - name: Dry run
-        run: source ./apps/ci/ci-worldserver-dry-run.sh
-      - name: Check startup errors
-        run: source ./apps/ci/ci-error-check.sh
+    uses: azerothcore/reusable-workflows/.github/workflows/core_build_modules.yml@main
+    with:
+      module_repo: ${{ github.event.repository.name }}
