+++ conf/player_not_speak.conf.dist
@@ -3,7 +3,7 @@
 #
 #    PlayerNotSpeak.Enable
 #        Description: Enable notification on chat if player receiver is muted.
-#        	  Option: 1  - (Enable)
+#        Option: 1  - (Enable)
 #                     0  - (Disable)
 
 PlayerNotSpeak.Enable = 1
