+++ data/sql/db-world/mnm_acore_string.sql
@@ -1,0 +1,3 @@
+DELETE FROM `acore_string` WHERE `entry`=13000;
+INSERT INTO `acore_string`(`entry`, `content_default`, `locale_koKR`, `locale_frFR`, `locale_deDE`, `locale_zhCN`, `locale_zhTW`, `locale_esES`, `locale_esMX`, `locale_ruRU`) VALUES
+(13000, 'Player %s will not be able to speak yet: %s', NULL, NULL, NULL, NULL, NULL, 'La jugador %s no podrá hablar todavía: %s', 'La jugador %s no podrá hablar todavía: %s', '|cffff0000# |cff00ff00Игрок|r %s |cff00ff00не сможет говорить ещё: |cffff0000%s');
