--- sql/world/trinity_string.sql
@@ -1,4 +1,0 @@
--- Language
-
-INSERT INTO `acore_string`(`entry`, `content_default`, `locale_ruRU`) VALUES
-(13000, 'Player %s will not be able to speak yet: %s', '|cffff0000# |cff00ff00Игрок|r %s |cff00ff00не сможет говорить ещё: |cffff0000%s');
