+++ src/mnm_loader.cpp
@@ -1,0 +1,6 @@
+void AddPlayerNotSpeakScripts();
+
+void Addmod_notify_mutedScripts()
+{
+    AddPlayerNotSpeakScripts();
+}
